backend/requirements.txt
alembic==1.13.3
axios==0.4.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-DebugToolbar==0.16.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
psycopg2==2.9.9
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.2
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
